feat: redesign all modal to main component
