+++ src/App.vue
@@ -23,8 +23,8 @@ export default {
 
 html,
 body {
-  // background-color: #fffcf7 !important;
-  background-image: url('../src/assets/images/blush-bg.jpg');
+  background-color: #fffcf7 !important;
+  // background-image: url('../src/assets/images/blush-bg.jpg');
   -webkit-background-size: cover;
   -moz-background-size: cover;
   -o-background-size: cover;
