+++ src/store/index.js
@@ -3,6 +3,7 @@ import Vuex from 'vuex';
 
 import user from '@/store/user.js';
 import makeupByRef from '@/store/makeupByRef.js';
+import uploadImage from '@/store/uploadImage.js';
 
 Vue.use(Vuex);
 
@@ -10,5 +11,6 @@ export default new Vuex.Store({
   modules: {
     user,
     makeupByRef,
+    uploadImage,
   },
 });
